+++ crates/burn-wgpu/src/codegen/dialect/wgsl/compiler.rs
@@ -313,6 +313,16 @@ impl<F: FloatElement, I: IntElement> Compiler<F, I> {
 
     fn compile_instruction(&mut self, value: gpu::Operator) -> wgsl::Instruction {
         match value {
+            gpu::Operator::Max(op) => wgsl::Instruction::Max {
+                lhs: self.compile_variable(op.lhs),
+                rhs: self.compile_variable(op.rhs),
+                out: self.compile_variable(op.out),
+            },
+            gpu::Operator::Min(op) => wgsl::Instruction::Min {
+                lhs: self.compile_variable(op.lhs),
+                rhs: self.compile_variable(op.rhs),
+                out: self.compile_variable(op.out),
+            },
             gpu::Operator::Add(op) => wgsl::Instruction::Add {
                 lhs: self.compile_variable(op.lhs),
                 rhs: self.compile_variable(op.rhs),
