+++ crates/burn-wgpu/src/fusion/tracing/builder.rs
@@ -186,6 +186,17 @@ impl TraceBuilder {
                         &mut local_tensor_ids_input,
                         &mut local_tensor_ids_output,
                     ),
+                    gpu::Operator::Max(op) => mark_binary(
+                        op,
+                        &mut local_tensor_ids_input,
+                        &mut local_tensor_ids_output,
+                    ),
+
+                    gpu::Operator::Min(op) => mark_binary(
+                        op,
+                        &mut local_tensor_ids_input,
+                        &mut local_tensor_ids_output,
+                    ),
                     gpu::Operator::And(op) => mark_binary(
                         op,
                         &mut local_tensor_ids_input,
