+++ crates/burn-wgpu/src/kernel/pool/mod.rs
@@ -2,10 +2,11 @@ mod adaptive_avg_pool2d;
 mod avg_pool2d;
 mod base;
 mod max_pool2d;
-mod max_pool2d_new;
+mod max_pool2d_backward;
 
 pub(crate) use adaptive_avg_pool2d::*;
 pub use avg_pool2d::*;
 pub(super) use base::*;
-pub use max_pool2d::*;
-pub use max_pool2d_new::*;
+
+pub(crate) use max_pool2d::*;
+pub(crate) use max_pool2d_backward::*;
